asb demo progress
