+++ platform/aspire/index.md
@@ -4,4 +4,5 @@ summary: Describes how to orchestrate the Particular Platform via the Aspire App
 reviewed: 2026-05-20
 related:
 - samples/hosting/aspire
+- samples/aspire/platform-asb
 ---
