--- samples/aspire/asb-transport/Core_10/Sales/PlaceOrderHandler.cs
@@ -1,32 +1,0 @@
-﻿using Messages;
-using Microsoft.Extensions.Logging;
-
-namespace Sales;
-
-public class PlaceOrderHandler(ILogger<PlaceOrderHandler> log) : IHandleMessages<PlaceOrder>
-{
-    public Task Handle(PlaceOrder message, IMessageHandlerContext context)
-    {
-        log.LogInformation("Received PlaceOrder, OrderId = {OrderId}", message.OrderId);
-
-        // This is normally where some business logic would occur
-
-        // Force a transient exception to demonstrate failures in telemetry data
-        var ran = Random.Shared.Next(0, 5) ;
-        if (ran == 0 || ran == 1 || ran == 5)
-        {
-            throw new Exception("Oops");
-        }
-
-        var publishOptions = new PublishOptions();
-        // https://docs.particular.net/nservicebus/operations/opentelemetry#traces-span-relationships-publish-operations
-        publishOptions.ContinueExistingTraceOnReceive();
-
-        var orderPlaced = new OrderPlaced
-        {
-            OrderId = message.OrderId
-        };
-
-        return context.Publish(orderPlaced, publishOptions);
-    }
-}
