+++ samples/aspire/platform-asb/Core_10/AspireDemo.AppHost/AspireDemo.AppHost.csproj
@@ -10,10 +10,8 @@
   </PropertyGroup>
 
   <ItemGroup>
-    <ProjectReference Include="..\Billing\Billing.csproj" />
     <ProjectReference Include="..\ClientUI\ClientUI.csproj" />
     <ProjectReference Include="..\Sales\Sales.csproj" />
-    <ProjectReference Include="..\Shipping\Shipping.csproj" />
   </ItemGroup>
 
   <ItemGroup>
