+++ samples/aspire/platform-asb/Core_10/AspireDemo.AppHost/appsettings.Development.json
