+++ samples/aspire/platform-asb/Core_10/AspireDemo.ServiceDefaults/AspireDemo.ServiceDefaults.csproj
@@ -19,7 +19,7 @@
     <PackageReference Include="NServiceBus.Heartbeat" Version="6.*" />
     <PackageReference Include="NServiceBus.Metrics.ServiceControl" Version="6.*" />
     <PackageReference Include="NServiceBus.Persistence.Sql" Version="9.*" />
-    <PackageReference Include="NServiceBus.RabbitMQ" Version="11.*" />
+    <PackageReference Include="NServiceBus.Transport.AzureServiceBus" Version="6.*" />
     <PackageReference Include="OpenTelemetry.Exporter.OpenTelemetryProtocol" Version="1.*" />
     <PackageReference Include="OpenTelemetry.Extensions.Hosting" Version="1.*" />
     <PackageReference Include="OpenTelemetry.Instrumentation.AspNetCore" Version="1.*" />
