+++ samples/aspire/platform-asb/Core_10/AspireDemo.ServiceDefaults/Extensions.cs
