+++ samples/aspire/platform-asb/Core_10/AspireDemo.ServiceDefaults/NServiceBusDefaults.cs
@@ -1,6 +1,4 @@
 using Microsoft.Extensions.Configuration;
-using NServiceBus;
-using NServiceBus.Transport;
 
 namespace Microsoft.Extensions.Hosting;
 
@@ -11,15 +9,12 @@ public static class NServiceBusDefaults
         public IHostApplicationBuilder AddNServiceBusEndpoint(string name,
             Action<EndpointConfiguration, RoutingSettings>? configureEndpoint = null)
         {
-            #region always-config
             var endpointConfiguration = new EndpointConfiguration(name);
-            #endregion
 
             #region transport-config
             var connectionString = builder.Configuration.GetConnectionString("transport")
-                    ?? throw new InvalidOperationException($"Endpoint '{name}' has no transport configured. Provide a 'ConnectionStrings:transport' connection string or set 'LEARNING_TRANSPORT_PATH'.");
-            var transport = new RabbitMQTransport(RoutingTopology.Conventional(QueueType.Quorum), connectionString);
-            var routing = endpointConfiguration.UseTransport(transport);
+                    ?? throw new InvalidOperationException($"Endpoint '{name}' has no transport configured. Provide a 'ConnectionStrings:transport'.");
+            var routing = endpointConfiguration.UseTransport(new AzureServiceBusTransport(connectionString, TopicTopology.Default));
             #endregion
 
             endpointConfiguration.UseSerialization<SystemJsonSerializer>();
