+++ samples/aspire/platform-asb/Core_10/Billing/Billing.csproj
@@ -10,7 +10,6 @@
 
   <ItemGroup>
     <ProjectReference Include="..\AspireDemo.ServiceDefaults\AspireDemo.ServiceDefaults.csproj" />
-    <ProjectReference Include="..\Messages\Messages.csproj" />
   </ItemGroup>
 
 </Project>
