+++ samples/aspire/platform-asb/Core_10/Billing/OrderPlacedHandler.cs
