+++ samples/aspire/platform-asb/Core_10/Messages/ShipOrder.cs
