+++ samples/aspire/platform-asb/Core_10/Sales.Messages/PlaceOrder.cs
@@ -1,0 +1,8 @@
+﻿using NServiceBus;
+
+namespace Messages;
+
+public class PlaceOrder : ICommand
+{
+    public string? OrderId { get; set; }
+}
