+++ samples/aspire/platform-asb/Core_10/Sales.Messages/Sales.Messages.csproj
@@ -1,0 +1,13 @@
+<Project Sdk="Microsoft.NET.Sdk">
+
+  <PropertyGroup>
+    <TargetFramework>net10.0</TargetFramework>
+    <ImplicitUsings>enable</ImplicitUsings>
+    <Nullable>enable</Nullable>
+  </PropertyGroup>
+
+  <ItemGroup>
+    <PackageReference Include="NServiceBus.MessageInterfaces" Version="1.*" />
+  </ItemGroup>
+
+</Project>
