+++ samples/aspire/platform-asb/Core_10/Sales/PlaceOrderHandler.cs
@@ -1,0 +1,18 @@
+﻿using Messages;
+using Microsoft.Extensions.Logging;
+
+namespace Sales;
+
+public class PlaceOrderHandler(ILogger<PlaceOrderHandler> logger) : IHandleMessages<PlaceOrder>
+{
+    public Task Handle(PlaceOrder message, IMessageHandlerContext context)
+    {
+        if (Random.Shared.Next(0, 5) == 0)
+        {
+            throw new Exception("Oops");
+        }
+
+        logger.LogInformation("Received PlaceOrder, OrderId = {OrderId}", message.OrderId);
+        return Task.CompletedTask;
+    }
+}
