+++ samples/aspire/platform-asb/Core_10/Shipping/Shipping.csproj
