+++ samples/aspire/platform-asb/Core_10/Shipping/ShippingPolicy.cs
